Fulfilled some of the tasks in pull requests like: add annotation @Repository to all repositories change all Float variables to double delete some of don't needed dependencies

+++ src/main/java/com/example/cashmanagerbackend/models/RegularIncome.java
@@ -28,7 +28,7 @@ public class RegularIncome {
     private String description;
 
     @Column(name = "cost", nullable = false)
-    private Float cost;
+    private Double cost;
 
     @Column(name = "last_payment_date", nullable = false)
     private OffsetDateTime lastPaymentDate;
